+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+* PR #1057: Enable async writes for greater throughput
 * PR #1094: CVE-2023-5072: Upgrading the org.json:json dependency
 * PR #1095: CVE-2023-4586: Upgrading the netty dependencies
 * BigQuery API has been upgraded to version 2.33.2