+++ CHANGES.md
@@ -5,6 +5,7 @@
 * PR #1057: Enable async writes for greater throughput
 * PR #1094: CVE-2023-5072: Upgrading the org.json:json dependency
 * PR #1095: CVE-2023-4586: Upgrading the netty dependencies
+* PR #1104: Fixed nested field predicate pushdown
 * Issue #1099: Fixing the usage of ExternalAccountCredentials
 * BigQuery API has been upgraded to version 2.33.2
 * BigQuery Storage API has been upgraded to version 2.44.0